Rename repo root scripts to use snake case
Snake_case is a common convention for shell scripts. ts scripts have
also been renamed to use snake_case to denote that they are scripts, and
not production code.

I'm unsure if we want to apply this convention in the hardhat subproject
at this point, and will leave that as-is for now.

Additionally, I also added the backend files to the typecheck yarn
script. This was an initial oversight, but has since been corrected. The
unused tsconfig.esrun.json has been deleted as well.

test plan:
`yarn db-seed-fresh`
`yarn generate`
`yarn clean`